+++ sqlalchemy_continuum/__init__.py
@@ -16,7 +16,7 @@ from .utils import (
 )
 
 
-__version__ = '1.0.2'
+__version__ = '1.0.3'
 
 
 versioning_manager = VersioningManager()